requirements.txt
Django==2.0.7
djangorestframework==3.8.2
python-decouple==3.1
pytz==2018.5
